refactor: move min spanning tree and bridges finder algorithms to separate classes
