+++ app/src/main/kotlin/model/graphs/UndirectedGraph.kt
@@ -61,57 +61,4 @@ open class UndirectedGraph<D> : Graph<D>() {
 
         return edge
     }
-
-    fun findBridges(): List<Edge<D>> {
-        val bridges = mutableListOf<Edge<D>>()
-
-        val graphSize = vertices.size
-
-        val discoveryTime = MutableList(graphSize) { -1 }
-        val minDiscoveryTime = MutableList(graphSize) { -1 }
-
-        val visitedList = MutableList(graphSize) { false }
-        val parentList = MutableList(graphSize) { -1 }
-
-        var iterationCount = 0
-
-        fun doDFSToFindBridgesFromVertex(vertex: Vertex<D>) {
-            visitedList[vertex.id] = true
-
-            iterationCount++
-            discoveryTime[vertex.id] = iterationCount
-            minDiscoveryTime[vertex.id] = iterationCount
-
-            for (neighbour in getNeighbours(vertex)) {
-                if (neighbour.id == parentList[vertex.id]) continue
-
-                if (visitedList[neighbour.id]) {
-                    minDiscoveryTime[vertex.id] =
-                        min(minDiscoveryTime[vertex.id], discoveryTime[neighbour.id])
-
-                    continue
-                }
-
-                parentList[neighbour.id] = vertex.id
-
-                doDFSToFindBridgesFromVertex(neighbour)
-
-                minDiscoveryTime[vertex.id] =
-                    min(minDiscoveryTime[vertex.id], minDiscoveryTime[neighbour.id])
-
-                if (minDiscoveryTime[neighbour.id] > discoveryTime[vertex.id]) {
-                    val bridgeFound = getEdge(vertex, neighbour)
-                    bridges.add(bridgeFound)
-                }
-            }
-        }
-
-        for (v in vertices) {
-            if (visitedList[v.id]) continue
-
-            doDFSToFindBridgesFromVertex(v)
-        }
-
-        return bridges
-    }
 }
